Add sprite attachment functionality to odd sprite channels
This commit adds the ability to attach sprites on odd sprite channels. A new function `spriteSetAttached()` is added which sets whether a sprite is attached or not. The header of the sprite is updated accordingly in `spriteProcess()`. An error message is logged if an even channel index is passed to `spriteSetAttached()`.