Add sprite attachment functionality for 16 colour Sprites.
This commit adds the ability to attach sprites, which are only available on odd sprite channels. The new function `spriteSetAttached()` sets whether a sprite is attached or not. If an even channel is passed in, an error message will be logged and the attachment flag will be set to 0. The `spriteProcess()` function has also been updated to include the attachment flag when updating metadata.